More merge fix
